SHM: use u64 increment in shmEnsureTable
@bare rejects int + 1 assignments; the two-PDE table walk must stay u64.

+++ core/kernel/shm.dart
@@ -968,7 +968,7 @@ u64 shmEnsureTable(u64 s) {
         return u64(shmRetNoTable);
       }
     }
-    i = i + 1;
+    i = i + u64(1);
   }
   procSet(s, u64(procSlotShmPt), vmShmTable());
   return u64(0);
